requirements.txt
Flask==3.0.2
firebase-admin==6.6.0
googlemaps==4.10.0
requests==2.32.3
twilio==9.3.0
python-dotenv==1.0.1
